requirements.txt
streamlit
pandas
pdfplumber
PyMuPDF
openpyxl
Pillow
